requirements.txt
tiktoken==0.8.0
openai==1.59.5
chromadb==0.6.2
sentence-transformers==3.3.1
FlagEmbedding==1.3.3 
pymupdf4llm==0.0.17
tenacity==9.0.0
python-dotenv==1.0.1